+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## v0.2.2 - 2017-04-29
+### Added:
+ - Resumes: edit, create
+ - Vacancies: favorited, similar, statistics
+ - Artifacts: removePhoto, editPhoto, uploadPhoto, uploadPortfolio
+
+### Changed:
+ - Artifacts: photo renamed to photos
+ - Removed unnecessary complexity from the code
+
 ## v0.2.1 - 2017-04-24
 ### Changed:
  - inSearch renamed to setIsInSearch